requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.1
prettytable==0.7.2
pyecharts==1.7.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
redis==3.4.1
simplejson==3.17.0
six==1.14.0
Werkzeug==1.0.1
xlrd==1.2.0
